Automatically generate check constraints for boolean and enum attributes
And make enums map to TINYINT by default.

+++ hibernate-core/src/main/java/org/hibernate/metamodel/model/convert/internal/OrdinalEnumValueConverter.java
@@ -60,7 +60,7 @@ public class OrdinalEnumValueConverter<E extends Enum> implements EnumValueConve
 
 	@Override
 	public int getJdbcTypeCode() {
-		return Types.INTEGER;
+		return Types.TINYINT;
 	}
 
 	@Override
